PullRequest: graal/11836

+++ substratevm/src/com.oracle.svm.core/src/com/oracle/svm/core/heap/StoredContinuation.java
@@ -26,14 +26,10 @@ package com.oracle.svm.core.heap;
 
 import com.oracle.svm.core.annotate.Hybrid;
 
-/**
- * This class is used for variably-sized objects that store continuation stack frames.
- *
- * For object layout and other implementation details, see {@link StoredContinuationImpl}.
- */
-@Hybrid(componentType = byte.class)
+/** Execution state of a continuation, use via {@link StoredContinuationAccess}. */
+@Hybrid(componentType = long.class)
 public final class StoredContinuation {
-    /** Must be allocated via {@link StoredContinuationImpl}. */
+    /** Must be allocated via {@link StoredContinuationAccess}. */
     private StoredContinuation() {
     }
 }
